Migrate model subcommand to analyze
