+++ charts/modeljob/templates/job.yaml
@@ -158,12 +158,12 @@ spec:
               {{- if gt (int $gpuCount) 0}}
               nvidia.com/gpu: {{ .Values.gpuCount }}
               {{- end }}
-              {{ - if gt (int $gpuMemory) 0 }}
-              aliyun.com/gpu-mem: { { .Values.gpuMemory }}
-              {{ - end } }
-              {{ - if gt (int $gpuCore) 0 }}
+              {{- if gt (int $gpuMemory) 0 }}
+              aliyun.com/gpu-mem: {{ .Values.gpuMemory }}
+              {{- end }}
+              {{- if gt (int $gpuCore) 0 }}
               aliyun.com/gpu-core.percentage: {{ .Values.gpuCore }}
-              {{ - end }}
+              {{- end }}
           volumeMounts:
             {{- if .Values.dataset }}
             {{- range $pvcName, $mntPath := .Values.dataset}}
