+++ pkg/apis/model/analyze/benchmark_builder.go
@@ -1,10 +1,11 @@
-package model
+package analyze
 
 import (
 	"fmt"
+	"strings"
+
 	"github.com/kubeflow/arena/pkg/apis/types"
 	"github.com/kubeflow/arena/pkg/argsbuilder"
-	"strings"
 )
 
 type ModelBenchmarkArgsBuilder struct {
