+++ pkg/apis/model/analyze/evaluate_builder.go
@@ -1,34 +1,35 @@
-package model
+package analyze
 
 import (
 	"fmt"
+	"strings"
+
 	"github.com/kubeflow/arena/pkg/apis/types"
 	"github.com/kubeflow/arena/pkg/argsbuilder"
-	"strings"
 )
 
-type ModelProfileJobBuilder struct {
-	args      *types.ModelProfileArgs
+type ModelEvaluateJobBuilder struct {
+	args      *types.ModelEvaluateArgs
 	argValues map[string]interface{}
 	argsbuilder.ArgsBuilder
 }
 
-func NewModelProfileJobBuilder() *ModelProfileJobBuilder {
-	args := &types.ModelProfileArgs{
+func NewModelEvaluateJobBuilder() *ModelEvaluateJobBuilder {
+	args := &types.ModelEvaluateArgs{
 		CommonModelArgs: types.CommonModelArgs{
 			Image:     argsbuilder.DefaultModelJobImage,
 			Namespace: "default",
 		},
 	}
-	return &ModelProfileJobBuilder{
+	return &ModelEvaluateJobBuilder{
 		args:        args,
 		argValues:   map[string]interface{}{},
-		ArgsBuilder: argsbuilder.NewModelProfileArgsBuilder(args),
+		ArgsBuilder: argsbuilder.NewModelEvaluateArgsBuilder(args),
 	}
 }
 
 // Name is used to set job name,match option --name
-func (m *ModelProfileJobBuilder) Name(name string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) Name(name string) *ModelEvaluateJobBuilder {
 	if name != "" {
 		m.args.Name = name
 	}
@@ -36,7 +37,7 @@ func (m *ModelProfileJobBuilder) Name(name string) *ModelProfileJobBuilder {
 }
 
 // Namespace is used to set job namespace,match option --namespace
-func (m *ModelProfileJobBuilder) Namespace(namespace string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) Namespace(namespace string) *ModelEvaluateJobBuilder {
 	if namespace != "" {
 		m.args.Namespace = namespace
 	}
@@ -44,7 +45,7 @@ func (m *ModelProfileJobBuilder) Namespace(namespace string) *ModelProfileJobBui
 }
 
 // Shell is used to specify linux shell type
-func (m *ModelProfileJobBuilder) Shell(shell string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) Shell(shell string) *ModelEvaluateJobBuilder {
 	if shell != "" {
 		m.args.Shell = shell
 	}
@@ -52,13 +53,13 @@ func (m *ModelProfileJobBuilder) Shell(shell string) *ModelProfileJobBuilder {
 }
 
 // Command is used to set job command
-func (m *ModelProfileJobBuilder) Command(args []string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) Command(args []string) *ModelEvaluateJobBuilder {
 	m.args.Command = strings.Join(args, " ")
 	return m
 }
 
 // Image is used to set job image,match the option --image
-func (m *ModelProfileJobBuilder) Image(image string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) Image(image string) *ModelEvaluateJobBuilder {
 	if image != "" {
 		m.args.Image = image
 	}
@@ -66,7 +67,7 @@ func (m *ModelProfileJobBuilder) Image(image string) *ModelProfileJobBuilder {
 }
 
 // ImagePullPolicy is used to set image pull policy,match the option --image-pull-policy
-func (m *ModelProfileJobBuilder) ImagePullPolicy(policy string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) ImagePullPolicy(policy string) *ModelEvaluateJobBuilder {
 	if policy != "" {
 		m.args.ImagePullPolicy = policy
 	}
@@ -74,7 +75,7 @@ func (m *ModelProfileJobBuilder) ImagePullPolicy(policy string) *ModelProfileJob
 }
 
 // ImagePullSecrets is used to set image pull secrests,match option --image-pull-secret
-func (m *ModelProfileJobBuilder) ImagePullSecrets(secrets []string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) ImagePullSecrets(secrets []string) *ModelEvaluateJobBuilder {
 	if secrets != nil {
 		m.argValues["image-pull-secret"] = &secrets
 	}
@@ -82,7 +83,7 @@ func (m *ModelProfileJobBuilder) ImagePullSecrets(secrets []string) *ModelProfil
 }
 
 // GPUCount is used to set count of gpu for the job,match the option --gpus
-func (m *ModelProfileJobBuilder) GPUCount(count int) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) GPUCount(count int) *ModelEvaluateJobBuilder {
 	if count > 0 {
 		m.args.GPUCount = count
 	}
@@ -90,15 +91,15 @@ func (m *ModelProfileJobBuilder) GPUCount(count int) *ModelProfileJobBuilder {
 }
 
 // GPUMemory is used to set gpu memory for the job,match the option --gpumemory
-func (m *ModelProfileJobBuilder) GPUMemory(memory int) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) GPUMemory(memory int) *ModelEvaluateJobBuilder {
 	if memory > 0 {
 		m.args.GPUMemory = memory
 	}
 	return m
 }
 
-// GPUCore is used to set gpu core for the job,match the option --gpucore
-func (m *ModelProfileJobBuilder) GPUCore(core int) *ModelProfileJobBuilder {
+// GPUCore is used to set gpu core for the job,match the option --gpumemory
+func (m *ModelEvaluateJobBuilder) GPUCore(core int) *ModelEvaluateJobBuilder {
 	if core > 0 {
 		m.args.GPUCore = core
 	}
@@ -106,7 +107,7 @@ func (m *ModelProfileJobBuilder) GPUCore(core int) *ModelProfileJobBuilder {
 }
 
 // CPU assign cpu limits,match the option --cpu
-func (m *ModelProfileJobBuilder) CPU(cpu string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) CPU(cpu string) *ModelEvaluateJobBuilder {
 	if cpu != "" {
 		m.args.Cpu = cpu
 	}
@@ -114,7 +115,7 @@ func (m *ModelProfileJobBuilder) CPU(cpu string) *ModelProfileJobBuilder {
 }
 
 // Memory assign memory limits,match option --memory
-func (m *ModelProfileJobBuilder) Memory(memory string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) Memory(memory string) *ModelEvaluateJobBuilder {
 	if memory != "" {
 		m.args.Memory = memory
 	}
@@ -122,7 +123,7 @@ func (m *ModelProfileJobBuilder) Memory(memory string) *ModelProfileJobBuilder {
 }
 
 // Envs is used to set env of job containers,match option --env
-func (m *ModelProfileJobBuilder) Envs(envs map[string]string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) Envs(envs map[string]string) *ModelEvaluateJobBuilder {
 	if len(envs) != 0 {
 		envSlice := []string{}
 		for key, value := range envs {
@@ -134,13 +135,13 @@ func (m *ModelProfileJobBuilder) Envs(envs map[string]string) *ModelProfileJobBu
 }
 
 // Tolerations is used to set tolerations for tolerate nodes,match option --toleration
-func (m *ModelProfileJobBuilder) Tolerations(tolerations []string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) Tolerations(tolerations []string) *ModelEvaluateJobBuilder {
 	m.argValues["toleration"] = &tolerations
 	return m
 }
 
 // NodeSelectors is used to set node selectors for scheduling job,match option --selector
-func (m *ModelProfileJobBuilder) NodeSelectors(selectors map[string]string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) NodeSelectors(selectors map[string]string) *ModelEvaluateJobBuilder {
 	if len(selectors) != 0 {
 		selectorsSlice := []string{}
 		for key, value := range selectors {
@@ -152,7 +153,7 @@ func (m *ModelProfileJobBuilder) NodeSelectors(selectors map[string]string) *Mod
 }
 
 // Annotations is used to add annotations for job pods,match option --annotation
-func (m *ModelProfileJobBuilder) Annotations(annotations map[string]string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) Annotations(annotations map[string]string) *ModelEvaluateJobBuilder {
 	if len(annotations) != 0 {
 		s := []string{}
 		for key, value := range annotations {
@@ -164,7 +165,7 @@ func (m *ModelProfileJobBuilder) Annotations(annotations map[string]string) *Mod
 }
 
 // Labels is used to add labels for job
-func (m *ModelProfileJobBuilder) Labels(labels map[string]string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) Labels(labels map[string]string) *ModelEvaluateJobBuilder {
 	if len(labels) != 0 {
 		s := []string{}
 		for key, value := range labels {
@@ -176,7 +177,7 @@ func (m *ModelProfileJobBuilder) Labels(labels map[string]string) *ModelProfileJ
 }
 
 // Datas is used to mount k8s pvc to job pods,match option --data
-func (m *ModelProfileJobBuilder) Datas(volumes map[string]string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) Datas(volumes map[string]string) *ModelEvaluateJobBuilder {
 	if len(volumes) != 0 {
 		s := []string{}
 		for key, value := range volumes {
@@ -188,7 +189,7 @@ func (m *ModelProfileJobBuilder) Datas(volumes map[string]string) *ModelProfileJ
 }
 
 // DataDirs is used to mount host files to job containers,match option --data-dir
-func (m *ModelProfileJobBuilder) DataDirs(volumes map[string]string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) DataDirs(volumes map[string]string) *ModelEvaluateJobBuilder {
 	if len(volumes) != 0 {
 		s := []string{}
 		for key, value := range volumes {
@@ -199,48 +200,64 @@ func (m *ModelProfileJobBuilder) DataDirs(volumes map[string]string) *ModelProfi
 	return m
 }
 
-// ModelConfigFile is used to set model config file,match the option --model-config-file
-func (m *ModelProfileJobBuilder) ModelConfigFile(filePath string) *ModelProfileJobBuilder {
-	if filePath != "" {
-		m.args.ModelConfigFile = filePath
+// SyncImage is used to set syncing image,match option --sync-image
+func (m *ModelEvaluateJobBuilder) SyncImage(image string) *ModelEvaluateJobBuilder {
+	if image != "" {
+		m.args.SyncImage = image
 	}
 	return m
 }
 
-// ModelName is used to set model name,match the option --model-name
-func (m *ModelProfileJobBuilder) ModelName(name string) *ModelProfileJobBuilder {
-	if name != "" {
-		m.args.ModelName = name
+// SyncMode is used to set syncing mode,match option --sync-mode
+func (m *ModelEvaluateJobBuilder) SyncMode(mode string) *ModelEvaluateJobBuilder {
+	if mode != "" {
+		m.args.SyncMode = mode
+	}
+	return m
+}
+
+// SyncSource is used to set syncing source,match option --sync-source
+func (m *ModelEvaluateJobBuilder) SyncSource(source string) *ModelEvaluateJobBuilder {
+	if source != "" {
+		m.args.SyncSource = source
 	}
 	return m
 }
 
 // ModelPath is used to set model path,match the option --model-path
-func (m *ModelProfileJobBuilder) ModelPath(path string) *ModelProfileJobBuilder {
+func (m *ModelEvaluateJobBuilder) ModelPath(path string) *ModelEvaluateJobBuilder {
 	if path != "" {
 		m.args.ModelPath = path
 	}
 	return m
 }
 
-// Inputs is used to specify model inputs
-func (m *ModelProfileJobBuilder) Inputs(inputs string) *ModelProfileJobBuilder {
-	if inputs != "" {
-		m.args.Inputs = inputs
+// ModelPlatform specify the model platform, such as torchscript/tensorflow
+func (m *ModelEvaluateJobBuilder) ModelPlatform(modelPlatform string) *ModelEvaluateJobBuilder {
+	if modelPlatform != "" {
+		m.args.ModelPlatform = modelPlatform
 	}
 	return m
 }
 
-// Outputs is used to specify model outputs
-func (m *ModelProfileJobBuilder) Outputs(outputs string) *ModelProfileJobBuilder {
-	if outputs != "" {
-		m.args.Outputs = outputs
+// DatasetPath is the dataset to evaluate model
+func (m *ModelEvaluateJobBuilder) DatasetPath(datasetPath string) *ModelEvaluateJobBuilder {
+	if datasetPath != "" {
+		m.args.DatasetPath = datasetPath
 	}
 	return m
 }
 
-// ReportPath is used to specify profile result path
-func (m *ModelProfileJobBuilder) ReportPath(reportPath string) *ModelProfileJobBuilder {
+// BatchSize is the batch size of evaluate
+func (m *ModelEvaluateJobBuilder) BatchSize(batchSize int) *ModelEvaluateJobBuilder {
+	if batchSize > 0 {
+		m.args.BatchSize = batchSize
+	}
+	return m
+}
+
+// ReportPath is used to specify evaluate result path
+func (m *ModelEvaluateJobBuilder) ReportPath(reportPath string) *ModelEvaluateJobBuilder {
 	if reportPath != "" {
 		m.args.ReportPath = reportPath
 	}
@@ -248,7 +265,7 @@ func (m *ModelProfileJobBuilder) ReportPath(reportPath string) *ModelProfileJobB
 }
 
 // Build is used to build the job
-func (m *ModelProfileJobBuilder) Build() (*Job, error) {
+func (m *ModelEvaluateJobBuilder) Build() (*Job, error) {
 	for key, value := range m.argValues {
 		m.AddArgValue(key, value)
 	}
@@ -258,5 +275,5 @@ func (m *ModelProfileJobBuilder) Build() (*Job, error) {
 	if err := m.ArgsBuilder.Build(); err != nil {
 		return nil, err
 	}
-	return NewJob(m.args.Name, types.ModelProfileJob, m.args), nil
+	return NewJob(m.args.Name, types.ModelEvaluateJob, m.args), nil
 }
