+++ pkg/apis/model/analyze/profile_builder.go
@@ -1,34 +1,35 @@
-package model
+package analyze
 
 import (
 	"fmt"
+	"strings"
+
 	"github.com/kubeflow/arena/pkg/apis/types"
 	"github.com/kubeflow/arena/pkg/argsbuilder"
-	"strings"
 )
 
-type ModelOptimizeJobBuilder struct {
-	args      *types.ModelOptimizeArgs
+type ModelProfileJobBuilder struct {
+	args      *types.ModelProfileArgs
 	argValues map[string]interface{}
 	argsbuilder.ArgsBuilder
 }
 
-func NewModelOptimizeJobBuilder() *ModelOptimizeJobBuilder {
-	args := &types.ModelOptimizeArgs{
+func NewModelProfileJobBuilder() *ModelProfileJobBuilder {
+	args := &types.ModelProfileArgs{
 		CommonModelArgs: types.CommonModelArgs{
 			Image:     argsbuilder.DefaultModelJobImage,
 			Namespace: "default",
 		},
 	}
-	return &ModelOptimizeJobBuilder{
+	return &ModelProfileJobBuilder{
 		args:        args,
 		argValues:   map[string]interface{}{},
-		ArgsBuilder: argsbuilder.NewModelOptimizeArgsBuilder(args),
+		ArgsBuilder: argsbuilder.NewModelProfileArgsBuilder(args),
 	}
 }
 
 // Name is used to set job name,match option --name
-func (m *ModelOptimizeJobBuilder) Name(name string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) Name(name string) *ModelProfileJobBuilder {
 	if name != "" {
 		m.args.Name = name
 	}
@@ -36,7 +37,7 @@ func (m *ModelOptimizeJobBuilder) Name(name string) *ModelOptimizeJobBuilder {
 }
 
 // Namespace is used to set job namespace,match option --namespace
-func (m *ModelOptimizeJobBuilder) Namespace(namespace string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) Namespace(namespace string) *ModelProfileJobBuilder {
 	if namespace != "" {
 		m.args.Namespace = namespace
 	}
@@ -44,7 +45,7 @@ func (m *ModelOptimizeJobBuilder) Namespace(namespace string) *ModelOptimizeJobB
 }
 
 // Shell is used to specify linux shell type
-func (m *ModelOptimizeJobBuilder) Shell(shell string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) Shell(shell string) *ModelProfileJobBuilder {
 	if shell != "" {
 		m.args.Shell = shell
 	}
@@ -52,13 +53,13 @@ func (m *ModelOptimizeJobBuilder) Shell(shell string) *ModelOptimizeJobBuilder {
 }
 
 // Command is used to set job command
-func (m *ModelOptimizeJobBuilder) Command(args []string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) Command(args []string) *ModelProfileJobBuilder {
 	m.args.Command = strings.Join(args, " ")
 	return m
 }
 
 // Image is used to set job image,match the option --image
-func (m *ModelOptimizeJobBuilder) Image(image string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) Image(image string) *ModelProfileJobBuilder {
 	if image != "" {
 		m.args.Image = image
 	}
@@ -66,7 +67,7 @@ func (m *ModelOptimizeJobBuilder) Image(image string) *ModelOptimizeJobBuilder {
 }
 
 // ImagePullPolicy is used to set image pull policy,match the option --image-pull-policy
-func (m *ModelOptimizeJobBuilder) ImagePullPolicy(policy string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) ImagePullPolicy(policy string) *ModelProfileJobBuilder {
 	if policy != "" {
 		m.args.ImagePullPolicy = policy
 	}
@@ -74,7 +75,7 @@ func (m *ModelOptimizeJobBuilder) ImagePullPolicy(policy string) *ModelOptimizeJ
 }
 
 // ImagePullSecrets is used to set image pull secrests,match option --image-pull-secret
-func (m *ModelOptimizeJobBuilder) ImagePullSecrets(secrets []string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) ImagePullSecrets(secrets []string) *ModelProfileJobBuilder {
 	if secrets != nil {
 		m.argValues["image-pull-secret"] = &secrets
 	}
@@ -82,7 +83,7 @@ func (m *ModelOptimizeJobBuilder) ImagePullSecrets(secrets []string) *ModelOptim
 }
 
 // GPUCount is used to set count of gpu for the job,match the option --gpus
-func (m *ModelOptimizeJobBuilder) GPUCount(count int) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) GPUCount(count int) *ModelProfileJobBuilder {
 	if count > 0 {
 		m.args.GPUCount = count
 	}
@@ -90,23 +91,23 @@ func (m *ModelOptimizeJobBuilder) GPUCount(count int) *ModelOptimizeJobBuilder {
 }
 
 // GPUMemory is used to set gpu memory for the job,match the option --gpumemory
-func (m *ModelOptimizeJobBuilder) GPUMemory(memory int) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) GPUMemory(memory int) *ModelProfileJobBuilder {
 	if memory > 0 {
 		m.args.GPUMemory = memory
 	}
 	return m
 }
 
-// GPUCore is used to set gpu core for the job, match the option --gpucore
-func (b *ModelOptimizeJobBuilder) GPUCore(core int) *ModelOptimizeJobBuilder {
+// GPUCore is used to set gpu core for the job,match the option --gpucore
+func (m *ModelProfileJobBuilder) GPUCore(core int) *ModelProfileJobBuilder {
 	if core > 0 {
-		b.args.GPUCore = core
+		m.args.GPUCore = core
 	}
-	return b
+	return m
 }
 
 // CPU assign cpu limits,match the option --cpu
-func (m *ModelOptimizeJobBuilder) CPU(cpu string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) CPU(cpu string) *ModelProfileJobBuilder {
 	if cpu != "" {
 		m.args.Cpu = cpu
 	}
@@ -114,7 +115,7 @@ func (m *ModelOptimizeJobBuilder) CPU(cpu string) *ModelOptimizeJobBuilder {
 }
 
 // Memory assign memory limits,match option --memory
-func (m *ModelOptimizeJobBuilder) Memory(memory string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) Memory(memory string) *ModelProfileJobBuilder {
 	if memory != "" {
 		m.args.Memory = memory
 	}
@@ -122,7 +123,7 @@ func (m *ModelOptimizeJobBuilder) Memory(memory string) *ModelOptimizeJobBuilder
 }
 
 // Envs is used to set env of job containers,match option --env
-func (m *ModelOptimizeJobBuilder) Envs(envs map[string]string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) Envs(envs map[string]string) *ModelProfileJobBuilder {
 	if len(envs) != 0 {
 		envSlice := []string{}
 		for key, value := range envs {
@@ -134,13 +135,13 @@ func (m *ModelOptimizeJobBuilder) Envs(envs map[string]string) *ModelOptimizeJob
 }
 
 // Tolerations is used to set tolerations for tolerate nodes,match option --toleration
-func (m *ModelOptimizeJobBuilder) Tolerations(tolerations []string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) Tolerations(tolerations []string) *ModelProfileJobBuilder {
 	m.argValues["toleration"] = &tolerations
 	return m
 }
 
 // NodeSelectors is used to set node selectors for scheduling job,match option --selector
-func (m *ModelOptimizeJobBuilder) NodeSelectors(selectors map[string]string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) NodeSelectors(selectors map[string]string) *ModelProfileJobBuilder {
 	if len(selectors) != 0 {
 		selectorsSlice := []string{}
 		for key, value := range selectors {
@@ -152,7 +153,7 @@ func (m *ModelOptimizeJobBuilder) NodeSelectors(selectors map[string]string) *Mo
 }
 
 // Annotations is used to add annotations for job pods,match option --annotation
-func (m *ModelOptimizeJobBuilder) Annotations(annotations map[string]string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) Annotations(annotations map[string]string) *ModelProfileJobBuilder {
 	if len(annotations) != 0 {
 		s := []string{}
 		for key, value := range annotations {
@@ -164,7 +165,7 @@ func (m *ModelOptimizeJobBuilder) Annotations(annotations map[string]string) *Mo
 }
 
 // Labels is used to add labels for job
-func (m *ModelOptimizeJobBuilder) Labels(labels map[string]string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) Labels(labels map[string]string) *ModelProfileJobBuilder {
 	if len(labels) != 0 {
 		s := []string{}
 		for key, value := range labels {
@@ -176,7 +177,7 @@ func (m *ModelOptimizeJobBuilder) Labels(labels map[string]string) *ModelOptimiz
 }
 
 // Datas is used to mount k8s pvc to job pods,match option --data
-func (m *ModelOptimizeJobBuilder) Datas(volumes map[string]string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) Datas(volumes map[string]string) *ModelProfileJobBuilder {
 	if len(volumes) != 0 {
 		s := []string{}
 		for key, value := range volumes {
@@ -188,7 +189,7 @@ func (m *ModelOptimizeJobBuilder) Datas(volumes map[string]string) *ModelOptimiz
 }
 
 // DataDirs is used to mount host files to job containers,match option --data-dir
-func (m *ModelOptimizeJobBuilder) DataDirs(volumes map[string]string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) DataDirs(volumes map[string]string) *ModelProfileJobBuilder {
 	if len(volumes) != 0 {
 		s := []string{}
 		for key, value := range volumes {
@@ -200,7 +201,7 @@ func (m *ModelOptimizeJobBuilder) DataDirs(volumes map[string]string) *ModelOpti
 }
 
 // ModelConfigFile is used to set model config file,match the option --model-config-file
-func (m *ModelOptimizeJobBuilder) ModelConfigFile(filePath string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) ModelConfigFile(filePath string) *ModelProfileJobBuilder {
 	if filePath != "" {
 		m.args.ModelConfigFile = filePath
 	}
@@ -208,7 +209,7 @@ func (m *ModelOptimizeJobBuilder) ModelConfigFile(filePath string) *ModelOptimiz
 }
 
 // ModelName is used to set model name,match the option --model-name
-func (m *ModelOptimizeJobBuilder) ModelName(name string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) ModelName(name string) *ModelProfileJobBuilder {
 	if name != "" {
 		m.args.ModelName = name
 	}
@@ -216,7 +217,7 @@ func (m *ModelOptimizeJobBuilder) ModelName(name string) *ModelOptimizeJobBuilde
 }
 
 // ModelPath is used to set model path,match the option --model-path
-func (m *ModelOptimizeJobBuilder) ModelPath(path string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) ModelPath(path string) *ModelProfileJobBuilder {
 	if path != "" {
 		m.args.ModelPath = path
 	}
@@ -224,7 +225,7 @@ func (m *ModelOptimizeJobBuilder) ModelPath(path string) *ModelOptimizeJobBuilde
 }
 
 // Inputs is used to specify model inputs
-func (m *ModelOptimizeJobBuilder) Inputs(inputs string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) Inputs(inputs string) *ModelProfileJobBuilder {
 	if inputs != "" {
 		m.args.Inputs = inputs
 	}
@@ -232,39 +233,23 @@ func (m *ModelOptimizeJobBuilder) Inputs(inputs string) *ModelOptimizeJobBuilder
 }
 
 // Outputs is used to specify model outputs
-func (m *ModelOptimizeJobBuilder) Outputs(outputs string) *ModelOptimizeJobBuilder {
+func (m *ModelProfileJobBuilder) Outputs(outputs string) *ModelProfileJobBuilder {
 	if outputs != "" {
 		m.args.Outputs = outputs
 	}
 	return m
 }
 
-// Optimizer is used to specify optimized model save path
-func (m *ModelOptimizeJobBuilder) Optimizer(optimizer string) *ModelOptimizeJobBuilder {
-	if optimizer != "" {
-		m.args.Optimizer = optimizer
-	}
-	return m
-}
-
-// TargetDevice is used to specify model deploy device
-func (m *ModelOptimizeJobBuilder) TargetDevice(targetDevice string) *ModelOptimizeJobBuilder {
-	if targetDevice != "" {
-		m.args.TargetDevice = targetDevice
-	}
-	return m
-}
-
-// ExportPath is used to specify optimized model save path
-func (m *ModelOptimizeJobBuilder) ExportPath(exportPath string) *ModelOptimizeJobBuilder {
-	if exportPath != "" {
-		m.args.ExportPath = exportPath
+// ReportPath is used to specify profile result path
+func (m *ModelProfileJobBuilder) ReportPath(reportPath string) *ModelProfileJobBuilder {
+	if reportPath != "" {
+		m.args.ReportPath = reportPath
 	}
 	return m
 }
 
 // Build is used to build the job
-func (m *ModelOptimizeJobBuilder) Build() (*Job, error) {
+func (m *ModelProfileJobBuilder) Build() (*Job, error) {
 	for key, value := range m.argValues {
 		m.AddArgValue(key, value)
 	}
@@ -274,5 +259,5 @@ func (m *ModelOptimizeJobBuilder) Build() (*Job, error) {
 	if err := m.ArgsBuilder.Build(); err != nil {
 		return nil, err
 	}
-	return NewJob(m.args.Name, types.ModelOptimizeJob, m.args), nil
+	return NewJob(m.args.Name, types.ModelProfileJob, m.args), nil
 }
