+++ pkg/commands/model/analyze/delete.go
@@ -1,24 +1,29 @@
-package model
+package analyze
 
 import (
 	"fmt"
+
 	"github.com/kubeflow/arena/pkg/apis/arenaclient"
-	"github.com/kubeflow/arena/pkg/apis/config"
-	"github.com/kubeflow/arena/pkg/apis/model"
 	"github.com/kubeflow/arena/pkg/apis/types"
+	"github.com/kubeflow/arena/pkg/apis/utils"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
-func NewSubmitModelProfileJobCommand() *cobra.Command {
-	builder := model.NewModelProfileJobBuilder()
+func NewDeleteModelJobCommand() *cobra.Command {
+	var jobType string
 	var command = &cobra.Command{
-		Use:   "profile",
-		Short: "Submit a model profile job",
+		Use:   "delete",
+		Short: "Delete a model analyze job",
 		PreRun: func(cmd *cobra.Command, args []string) {
 			_ = viper.BindPFlags(cmd.Flags())
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
+			if len(args) == 0 {
+				return fmt.Errorf("not set job name,please set it")
+			}
+			names := args
+
 			client, err := arenaclient.NewArenaClient(types.ArenaClientArgs{
 				Kubeconfig:     viper.GetString("config"),
 				LogLevel:       viper.GetString("loglevel"),
@@ -27,15 +32,12 @@ func NewSubmitModelProfileJobCommand() *cobra.Command {
 				IsDaemonMode:   false,
 			})
 			if err != nil {
-				return fmt.Errorf("failed to create arena client: %v\n", err)
+				return err
 			}
-			job, err := builder.Namespace(config.GetArenaConfiger().GetNamespace()).Command(args).Build()
-			if err != nil {
-				return fmt.Errorf("failed to validate command args: %v", err)
-			}
-			return client.Model().Submit(job)
+			return client.Analyze().Delete(utils.TransferModelJobType(jobType), names...)
 		},
 	}
-	builder.AddCommandFlags(command)
+
+	command.Flags().StringVarP(&jobType, "type", "T", "", fmt.Sprintf("The model job type to delete, the possible option is [%v]. (optional)", utils.GetSupportModelJobTypesInfo()))
 	return command
 }
