+++ pkg/commands/model/analyze/submit_benchmark.go
@@ -1,29 +1,25 @@
-package model
+package analyze
 
 import (
 	"fmt"
+
 	"github.com/kubeflow/arena/pkg/apis/arenaclient"
+	"github.com/kubeflow/arena/pkg/apis/config"
+	"github.com/kubeflow/arena/pkg/apis/model/analyze"
 	"github.com/kubeflow/arena/pkg/apis/types"
-	"github.com/kubeflow/arena/pkg/apis/utils"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
-func NewGetModelJobCommand() *cobra.Command {
-	var jobType string
-	var output string
+func NewSubmitModelBenchmarkJobCommand() *cobra.Command {
+	builder := analyze.NewModelBenchmarkArgsBuilder()
 	var command = &cobra.Command{
-		Use:   "get",
-		Short: "Get a model job",
+		Use:   "benchmark",
+		Short: "Submit a model benchmark job",
 		PreRun: func(cmd *cobra.Command, args []string) {
 			_ = viper.BindPFlags(cmd.Flags())
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if len(args) == 0 {
-				return fmt.Errorf("not set job name,please set it")
-			}
-			name := args[0]
-
 			client, err := arenaclient.NewArenaClient(types.ArenaClientArgs{
 				Kubeconfig:     viper.GetString("config"),
 				LogLevel:       viper.GetString("loglevel"),
@@ -32,14 +28,15 @@ func NewGetModelJobCommand() *cobra.Command {
 				IsDaemonMode:   false,
 			})
 			if err != nil {
-				return err
+				return fmt.Errorf("failed to create arena client: %v\n", err)
 			}
-			return client.Model().GetAndPrint(utils.TransferModelJobType(jobType), name, output)
+			job, err := builder.Namespace(config.GetArenaConfiger().GetNamespace()).Command(args).Build()
+			if err != nil {
+				return fmt.Errorf("failed to validate command args: %v", err)
+			}
+			return client.Analyze().Submit(job)
 		},
 	}
-
-	command.Flags().StringVarP(&jobType, "type", "T", "", fmt.Sprintf("The model job type to delete, the possible option is [%v]. (optional)", utils.GetSupportModelJobTypesInfo()))
-	command.Flags().StringVarP(&output, "output", "o", "wide", "Output format. One of: json|yaml|wide")
-
+	builder.AddCommandFlags(command)
 	return command
 }
