+++ pkg/commands/model/model.go
@@ -1,33 +1,19 @@
 package model
 
-import "github.com/spf13/cobra"
+import (
+	"github.com/kubeflow/arena/pkg/commands/model/analyze"
 
-var (
-	serveLong = `submit a model analyze job.
-
-Available Commands:
-  profile          Submit a model profile job.
-  evaluate         Submit a model evaluate job.
-  optimize         Submit a model optimize job.
-  benchmark        Submit a model benchmark job`
+	"github.com/spf13/cobra"
 )
 
 func NewModelCommand() *cobra.Command {
 	var command = &cobra.Command{
 		Use:   "model",
-		Short: "Submit a model analyze job. (experimental feature)",
-		Long:  serveLong,
+		Short: "Model manage and model analyze",
 		Run: func(cmd *cobra.Command, args []string) {
 			cmd.HelpFunc()(cmd, args)
 		},
 	}
-	command.AddCommand(NewSubmitModelProfileJobCommand())
-	command.AddCommand(NewSubmitModelOptimizeJobCommand())
-	command.AddCommand(NewSubmitModelBenchmarkJobCommand())
-	command.AddCommand(NewSubmitModelEvaluateJobCommand())
-	command.AddCommand(NewGetModelJobCommand())
-	command.AddCommand(NewListModelJobsCommand())
-	command.AddCommand(NewDeleteModelJobCommand())
-
+	command.AddCommand(analyze.NewAnalyzeCommand())
 	return command
 }
