+++ pkg/model/analyze/model_interface.go
@@ -1,11 +1,12 @@
-package model
+package analyze
 
 import (
+	"time"
+
 	"github.com/kubeflow/arena/pkg/apis/types"
 	batchv1 "k8s.io/api/batch/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"time"
 )
 
 type ModelJob interface {
